added function for each argument and working on error handling

+++ src/main.js
@@ -2,53 +2,68 @@ import chalk from 'chalk';
 import execa from 'execa';
 import fs, { exists } from 'fs';
 import Listr from 'listr';
+import figlet from "figlet"
 import ncp from 'ncp';
 import path from 'path';
-import { projectInstall } from 'pkg-install';
 import { promisify } from 'util';
 
 const access = promisify(fs.access);
-const copy = promisify(ncp);
+
 let exist = false;
-async function copyTemplateFiles(options) {
-  if(options.module){
+
+const createModule = (options) => {
+  if (options.module) {
     if (options.module_name && !options.controller && !options.service && !options.migration) {
-      copyDirectoryRecursiveSync(options.templateDirectory + '/' + options.module+'/Modules', options.targetDirectory + '/Modules');
-      copyDirectoryRecursiveSync(options.templateDirectory + '/' + options.module +'/config', options.targetDirectory + '/config');
-      copyDirectoryRecursiveSync(options.templateDirectory + '/' + options.module +'/functions', options.targetDirectory + '/functions');
-      
+      copyDirectoryRecursiveSync(options.templateDirectory + '/' + options.module + '/Modules', options.targetDirectory + '/Modules', true);
+      copyDirectoryRecursiveSync(options.templateDirectory + '/' + options.module + '/config', options.targetDirectory + '/config',false);
+      copyDirectoryRecursiveSync(options.templateDirectory + '/' + options.module + '/functions', options.targetDirectory + '/functions', false);
     }
+    if (exist === false) {
+      renameFilesRecursive(options.targetDirectory + '/Modules', /File/g, options.module_name);
+      renameFilesRecursive(options.targetDirectory + '/Modules', /file/g, options.module_name.toLowerCase());
+      writeFile(options.targetDirectory + '/Modules' + '/' + options.module_name, options.module_name.toLowerCase())
+    }
+  }
+}
+
+const createModuleController = (options) => {
+  if (options.module) {
     if (options.module_name && options.controller) {
-      copyDirectoryRecursiveSync(options.templateDirectory +'/'+ options.module +'/Modules/File/Controllers', options.targetDirectory + '/Modules/' + options.module_name + '/Controllers');
+      copyDirectoryRecursiveSync(options.templateDirectory + '/' + options.module + '/Modules/File/Controllers', options.targetDirectory + '/Modules/' + options.module_name + '/Controllers');
       renameFilesRecursive(options.targetDirectory + '/Modules/' + options.module_name + '/Controllers', /File/g, options.controller);
     }
+  }
+}
+
+const createModuleService = (options) => {
+  if (options.module) {
     if (options.module_name && options.service) {
       copyDirectoryRecursiveSync(options.templateDirectory + '/' + options.module + '/Modules/File/Service', options.targetDirectory + '/Modules/' + options.module_name + '/Service');
       renameFilesRecursive(options.targetDirectory + '/Modules/' + options.module_name + '/Service', /File/g, options.service);
-    } 
-    if (options.module_name && options.migration) {
-      if (options.module == " sequelize") {
-        copyDirectoryRecursiveSync(options.templateDirectory + '/' + options.module + '/Modules/File/migrations', options.targetDirectory + '/Modules/' + options.module_name + '/migrations');
-        renameFilesRecursive(options.targetDirectory + '/Modules/' + options.module_name + '/migrations', /File/g, options.migration);
-        copyDirectoryRecursiveSync(options.templateDirectory + '/' + options.module + '/Modules/File/models', options.targetDirectory + '/Modules/' + options.module_name + '/models');
-        renameFilesRecursive(options.targetDirectory + '/Modules/' + options.module_name + '/models', /File/g, options.migration);
-      }else{
-        copyDirectoryRecursiveSync(options.templateDirectory + '/' + options.module + '/Modules/File/Models', options.targetDirectory + '/Modules/' + options.module_name + '/Models');
-        renameFilesRecursive(options.targetDirectory + '/Modules/' + options.module_name + '/Models', /File/g, options.migration);
-      }
-    }else{
-      if(exist === false){
-        renameFilesRecursive(options.targetDirectory + '/Modules', /File/g, options.module_name);
-        renameFilesRecursive(options.targetDirectory + '/Modules', /file/g, options.module_name.toLowerCase());
-        writeFile(options.targetDirectory + '/Modules' + '/' + options.module_name, options.module_name.toLowerCase())
-      }
     }
-
   }
-  
 }
 
-async function copyDirectoryRecursiveSync(source, target, move) {
+const createModuleModelMigration = (options) => {
+  if (options.module) {
+    if (options.module_name && options.model_migration) {
+      if (options.module == " sequelize") {
+        copyDirectoryRecursiveSync(options.templateDirectory + '/' + options.module + '/Modules/File/migrations', options.targetDirectory + '/Modules/' + options.module_name + '/migrations');
+        renameFilesRecursive(options.targetDirectory + '/Modules/' + options.module_name + '/migrations', /File/g, Date.now() + '-create-'.options.model_migration.toLowerCase(), options.model_migration);
+        copyDirectoryRecursiveSync(options.templateDirectory + '/' + options.module + '/Modules/File/models', options.targetDirectory + '/Modules/' + options.module_name + '/models');
+        renameFilesRecursive(options.targetDirectory + '/Modules/' + options.module_name + '/models', /File/g, options.model_migration);
+      } else {
+        copyDirectoryRecursiveSync(options.templateDirectory + '/' + options.module + '/Modules/File/Models', options.targetDirectory + '/Modules/' + options.module_name + '/Models');
+        renameFilesRecursive(options.targetDirectory + '/Modules/' + options.module_name + '/Models', /File/g, options.model_migration);
+      }
+    }
+  }
+}
+
+async function copyDirectoryRecursiveSync(source, target,start = false, move) {
+  if(start === true){
+    init()
+  }
   if (!fs.lstatSync(source).isDirectory()) return
 
   if (!fs.existsSync(target)) {
@@ -69,40 +84,46 @@ async function copyDirectoryRecursiveSync(source, target, move) {
         copyDirectoryRecursiveSync(sourcePath, targetPath)
       } else {
         exist = true
-        console.log('it exist')
+        console.log('%s Module already exist', chalk.blue.bold('EXIST'))
+        process.exit()
       }
     }
     else {
       if (!fs.existsSync(targetPath)) {
-        console.log('1', targetPath)
+        // console.log('1', targetPath)
         operation(sourcePath, targetPath)
       }
     }
-  },function(err){
+  }, function (err) {
     console.log(err)
   })
 
 
 }
 
-async function renameFilesRecursive(dir, from, to) {
+async function renameFilesRecursive(dir, from, to, alt = '1') {
 
   // console.log(dir,from)
   fs.readdirSync(dir).forEach(it => {
     const itsPath = path.resolve(dir, it);
     const itsStat = fs.statSync(itsPath);
-    
+
     if (itsPath.search(from) > -1) {
       if (fs.existsSync(itsPath)) {
         try {
-          fs.renameSync(itsPath, itsPath.replace(from, to), function (error) {
-            if (error) { deleteFolderRecursive(error.path) }
-          })
-          FileReader(itsPath.replace(from, to), to)
-        }catch(error){
+          if (!fs.existsSync(itsPath.replace(from, to))) {
+            fs.renameSync(itsPath, itsPath.replace(from, to), function (error) {
+              if (error) { deleteFolderRecursive(error.path) }
+            })
+              (alt == '1') ? FileReader(itsPath.replace(from, to), to) : FileReader(itsPath.replace(from, to), alt)
+          } else {
+            console.log('%s ' + to + ' exist already', chalk.blue.bold('EXIST'))
+            process.exit()
+          }
+        } catch (error) {
           deleteFolderRecursive(error.path)
-        }        
-      }else{
+        }
+      } else {
         deleteFolderRecursive(error.path)
       }
     }
@@ -111,9 +132,9 @@ async function renameFilesRecursive(dir, from, to) {
       if (fs.existsSync(itsPath.replace(from, to))) {
         renameFilesRecursive(itsPath.replace(from, to), from, to)
       } else {
-        renameFilesRecursive(itsPath.replace(from, to), '/' + to + '/g', to)
+        renameFilesRecursive(itsPath.replace(to, from), '/' + to + '/g', to)
       }
-      
+
     }
   })
 }
@@ -126,7 +147,7 @@ async function FileReader(filePath, name) {
     var result = data.replace(/File/g, name);
 
     fs.writeFile(filePath, result, 'utf8', function (err) {
-      if (err) return console.log(err);
+      if (err) return console.log('%s '+err, chalk.red.bold('ERROR'));
     });
   });
 }
@@ -145,6 +166,18 @@ async function deleteFolderRecursive(path) {
   }
 };
 
+const init = () => {
+  console.log(
+    chalk.green(
+      figlet.textSync("Modular Express", {
+        font: "Ghost",
+        horizontalLayout: "default",
+        verticalLayout: "default"
+      })
+    )
+  );
+};
+
 async function writeFile(dir, name) {
   var fileName = dir + '/' + name + '.json';
   if (fs.existsSync(fileName)) {
@@ -152,9 +185,9 @@ async function writeFile(dir, name) {
     file.name = name + " Module";
     fs.writeFile(fileName, JSON.stringify(file, null, 2), function (err) {
       if (err) return console.log(err);
-      console.log(name.toUpperCase()+' Module');
+      console.log(name.toUpperCase() + ' Module');
     });
-  }else{
+  } else {
 
   }
 }
@@ -167,9 +200,9 @@ export async function createProject(options) {
     name: 'Bashorun Adedotun',
   };
 
-  
+
   let templateDir = path.resolve(__dirname);
-  templateDir = path.resolve(templateDir, templateDir,'../lib')
+  templateDir = path.resolve(templateDir, templateDir, '../lib')
   options.templateDirectory = templateDir;
   // console.log(options)
   try {
@@ -182,19 +215,8 @@ export async function createProject(options) {
   const tasks = new Listr(
     [
       {
-        title: 'Copy project files',
-        task: () => copyTemplateFiles(options),
-      },
-      {
-        title: 'Install dependencies',
-        task: () =>
-          projectInstall({
-            cwd: options.targetDirectory,
-          }),
-        skip: () =>
-          !options.runInstall
-            ? 'Pass --install to automatically install dependencies'
-            : undefined,
+        title: 'Create project files',
+        task: () => createModule(options),
       },
     ],
     {
@@ -204,5 +226,116 @@ export async function createProject(options) {
 
   await tasks.run();
   console.log('%s Project ready', chalk.green.bold('DONE'));
+  return true;
+}
+
+export async function createController(options) {
+  options = {
+    ...options,
+    targetDirectory: options.targetDirectory || process.cwd(),
+    email: 'adedotunolawale@gmail.com',
+    name: 'Bashorun Adedotun',
+  };
+
+
+  let templateDir = path.resolve(__dirname);
+  templateDir = path.resolve(templateDir, templateDir, '../lib')
+  options.templateDirectory = templateDir;
+  // console.log(options)
+  try {
+    await access(templateDir, fs.constants.R_OK);
+  } catch (err) {
+    console.error('%s Invalid modules name', chalk.red.bold('ERROR'));
+    process.exit(1);
+  }
+
+  const tasks = new Listr(
+    [
+      {
+        title: 'Creating Controller',
+        task: () => createModuleController(options),
+      },
+    ],
+    {
+      exitOnError: false,
+    }
+  );
+
+  await tasks.run();
+  console.log('%s Project ready', chalk.green.bold('DONE'));
+  return true;
+}
+
+export async function createService(options) {
+  options = {
+    ...options,
+    targetDirectory: options.targetDirectory || process.cwd(),
+    email: 'adedotunolawale@gmail.com',
+    name: 'Bashorun Adedotun',
+  };
+
+
+  let templateDir = path.resolve(__dirname);
+  templateDir = path.resolve(templateDir, templateDir, '../lib')
+  options.templateDirectory = templateDir;
+  // console.log(options)
+  try {
+    await access(templateDir, fs.constants.R_OK);
+  } catch (err) {
+    console.error('%s Invalid modules name', chalk.red.bold('ERROR'));
+    process.exit(1);
+  }
+
+  const tasks = new Listr(
+    [
+      {
+        title: 'Creating Service',
+        task: () => createModuleService(options),
+      },
+    ],
+    {
+      exitOnError: false,
+    }
+  );
+
+  await tasks.run();
+  console.log('%s Service ready', chalk.green.bold('DONE'));
+  return true;
+}
+
+export async function createModelMigration(options) {
+  options = {
+    ...options,
+    targetDirectory: options.targetDirectory || process.cwd(),
+    email: 'adedotunolawale@gmail.com',
+    name: 'Bashorun Adedotun',
+  };
+
+
+  let templateDir = path.resolve(__dirname);
+  templateDir = path.resolve(templateDir, templateDir, '../lib')
+  options.templateDirectory = templateDir;
+  // console.log(options)
+  try {
+    await access(templateDir, fs.constants.R_OK);
+  } catch (err) {
+    console.error('%s Invalid modules name', chalk.red.bold('ERROR'));
+    process.exit(1);
+  }
+
+  const tasks = new Listr(
+    [
+      {
+        title: 'Creating Model',
+        task: () => createModuleModelMigration(options),
+      },
+    ],
+    {
+      exitOnError: false,
+    }
+  );
+
+  await tasks.run();
+  console.log('%s Model ready', chalk.green.bold('DONE'));
   return true;
 }
